+++ src/node-addon/secur32-facade.cc
@@ -121,6 +121,13 @@ void Secur32Facade::FreeContextHandle(struct _SecHandle* ctxHandle, Napi::Env* e
   }
 }
 
+unsigned long Secur32Facade::GetDefaultFlags(std::string securityPackageName) {
+  if (securityPackageName == "Negotiate" || securityPackageName == "Kerberos") {
+    return ISC_REQ_MUTUAL_AUTH | ISC_REQ_SEQUENCE_DETECT;
+  }
+  return 0;
+}
+
 Napi::Number Secur32Facade::CreateAuthContext(const Napi::CallbackInfo& info) {
   Napi::Env env = info.Env();
   if (info.Length() < 4) {
@@ -128,17 +135,17 @@ Napi::Number Secur32Facade::CreateAuthContext(const Napi::CallbackInfo& info) {
     return Napi::Number::New(env, 0);
   }
 
-  if (!info[0].IsString() || !info[1].IsString() || !info[2].IsBuffer() || !info[3].IsBoolean()) {
+  if (!info[0].IsString() || !info[1].IsString() || !info[2].IsBuffer() || !(info[3].IsUndefined() || info[3].IsNumber())) {
     ExceptionHandler::CreateAndThrow(env, "Wrong argument types");
     return Napi::Number::New(env, 0);
   }
   auto securityPackageName = info[0].ToString();
   auto targetHost = info[1].ToString();
   auto applicationDataBuffer = info[2].As<Napi::Buffer<unsigned char>>();
-  auto delegate = info[3].ToBoolean();
+  auto flags = info[3].IsNumber() ? (unsigned long)(info[3].ToNumber().Uint32Value()) : GetDefaultFlags(securityPackageName);
 
-  auto ac = std::make_shared<AuthContext>(delegate);
-  ac->Init(&(securityPackageName.Utf8Value()), &(targetHost.Utf8Value()), applicationDataBuffer, &env);
+  auto ac = std::make_shared<AuthContext>();
+  ac->Init(&(securityPackageName.Utf8Value()), &(targetHost.Utf8Value()), applicationDataBuffer, flags, &env);
   acKey++;
   acMap[acKey] = ac;
   return Napi::Number::New(env, acKey);