+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.2.1 - released 2022-09-20
+
+* Make authentication context flags configurable
+
+## 1.2.0 - released 2022-09-19
+
+* Dependency bump
+* Improved Kerberos 
+  - delegation support
+  - set SPN earlier
+  - added security flags
+
 ## 1.1.8 - released 2022-04-23
 
 * Dependency bump